requirements.txt
asgiref==3.5.2
Django>=3.2,<4.0
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
drf-yasg
django-filter
python-dotenv==0.19.2
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.6
Pillow>=9.0.0
beautifulsoup4
requests
django_ratelimit
python-telegram-bot>=20.0,<21.0
openpyxl
redis
django-redis
django-cacheops 
django-mptt
pyyaml
ua-parser
user-agents
